Redoing the entire site
This is the first push of the major changes I am working on for the website. I wasn't happy with the look of the original, and with it being unfinished anyways I have decided to redo the whole design. This is still very incomplete, more so than before, but now I can't just say it's sorta done.